backend/requirements.txt
openai
Flask
python-dotenv
langchain